+++ ci/build-mac.sh
@@ -1,11 +1,19 @@
 #!/bin/bash
 
 VERSION=$(<VERSION)
-BASE_PATH=$(pwd)
 BUILD_PATH=$(pwd)/bin
 MODULE_PATH=$(pwd)/cmd/elk
 
+COMMIT=$(git rev-parse --short HEAD)
+VERSION=$(git describe --tags $(git rev-list --tags --max-count=1))
+
 GOOS=darwin
+
+day=$(date +'%a')
+month=$(date +'%b')
+fill_date=$(date +'%d_%T_%Y')
+
+DATE="${day^}_${month^}_${fill_date}"
 
 # Build for 386
 GOARCH=386
@@ -13,10 +21,10 @@ cd $MODULE_PATH
 NAME=elk
 
 BIN_PATH=$BUILD_PATH/$NAME
-go build -o $BIN_PATH
+go build -ldflags "-X main.v=$VERSION -X main.o=$GOOS -X main.arch=$GOARCH -X main.commit=$COMMIT -X main.date=$DATE" -o $BIN_PATH
 
 cd $BUILD_PATH
-ZIP_PATH=${BIN_PATH}_v${VERSION}_${GOOS}_${GOARCH}.zip
+ZIP_PATH=${BIN_PATH}_${VERSION}_${GOOS}_${GOARCH}.zip
 
 zip $ZIP_PATH $NAME
 rm $NAME
@@ -30,7 +38,7 @@ BIN_PATH=$BUILD_PATH/$NAME
 go build -o $BIN_PATH
 
 cd $BUILD_PATH
-ZIP_PATH=${BIN_PATH}_v${VERSION}_${GOOS}_${GOARCH}.zip
+ZIP_PATH=${BIN_PATH}_${VERSION}_${GOOS}_${GOARCH}.zip
 
 zip $ZIP_PATH $NAME
 rm $NAME