+++ internal/cli/command/command_test.go
@@ -6,7 +6,7 @@ import (
 )
 
 func TestExecute(t *testing.T) {
-	os.Args = []string{"elk"}
+	os.Args = []string{"ox"}
 	err := Execute()
 	if err != nil {
 		t.Error(err)