+++ internal/cli/command/cron/cmd.go
@@ -18,7 +18,8 @@ var usageTemplate = `Usage:
 
 Flags:
   -d, --detached         Run the task in detached mode and returns the PGID
-  -e, --env strings      Overwrite env variable in task   
+  -e, --env strings      Overwrite env variable in task
+  -v, --var strings      Overwrite var variable in task   
   -f, --file string      Run elk in a specific file
   -g, --global           Run from the path set in config
   -h, --help             Help for run
@@ -35,6 +36,7 @@ Flags:
 // NewRunCommand returns a cobra command for `run` sub command
 func NewCronCommand() *cobra.Command {
 	var envs []string
+	var vars []string
 	var cmd = &cobra.Command{
 		Use:   "cron",
 		Short: "Run one or more task as a cron job ⏱",
@@ -46,7 +48,7 @@ func NewCronCommand() *cobra.Command {
 				return
 			}
 
-			err = Run(cmd, args, envs)
+			err = Run(cmd, args, envs, vars)
 			if err != nil {
 				utils.PrintError(err)
 			}
@@ -55,6 +57,7 @@ func NewCronCommand() *cobra.Command {
 
 	cmd.Flags().BoolP("global", "g", false, "Run from the path set in config")
 	cmd.Flags().StringSliceVarP(&envs, "env", "e", []string{}, "")
+	cmd.Flags().StringSliceVarP(&vars, "var", "v", []string{}, "")
 	cmd.Flags().Bool("ignore-log-file", false, "Force task to output to stdout")
 	cmd.Flags().Bool("ignore-error", false, "Ignore errors that happened during a task")
 	cmd.Flags().BoolP("detached", "d", false, "Run the command in detached mode and returns the PGID")
@@ -70,7 +73,7 @@ func NewCronCommand() *cobra.Command {
 	return cmd
 }
 
-func Run(cmd *cobra.Command, args []string, envs []string) error {
+func Run(cmd *cobra.Command, args []string, envs []string, vars []string) error {
 	isDetached, err := cmd.Flags().GetBool("detached")
 	if err != nil {
 		return err
@@ -133,6 +136,12 @@ func Run(cmd *cobra.Command, args []string, envs []string) error {
 			env := strings.ReplaceAll(parts[0], "=", "")
 			value := parts[1]
 			task.Env[env] = value
+		}
+
+		for _, v := range vars {
+			parts := strings.SplitAfterN(v, "=", 2)
+			k := strings.ReplaceAll(parts[0], "=", "")
+			task.Vars[k] = parts[1]
 		}
 
 		clientEngine.Elk.Tasks[name] = task