+++ pkg/primitives/elk/error.go
@@ -1,7 +1,0 @@
-package elk
-
-import "errors"
-
-var ErrCircularDependency = errors.New("circular dependency")
-
-var ErrTaskNotFound = errors.New("task not found")