+++ pkg/primitives/ox/error.go
@@ -1,0 +1,7 @@
+package ox
+
+import "errors"
+
+var ErrCircularDependency = errors.New("circular dependency")
+
+var ErrTaskNotFound = errors.New("task not found")